images

+++ calcularNuevosPrecios-service/src/main/java/com/pep3/calcularNuevosPrecios/Model/Producto.java
@@ -4,9 +4,9 @@ public class Producto {
 
 	private int id;
 	private String nombre;
-	private int precio_peso;
-	private int precio_dolar;
-	private int precio_euro;
+	private float precio_peso;
+	private float precio_dolar;
+	private float precio_euro;
 	
 	public int getId() {
 		return id;
@@ -21,22 +21,22 @@ public class Producto {
 	public void setNombre(String nombre) {
 		this.nombre = nombre;
 	}
-	public int getPrecio_peso() {
+	public float getPrecio_peso() {
 		return precio_peso;
 	}
-	public void setPrecio_peso(int precio_peso) {
+	public void setPrecio_peso(float precio_peso) {
 		this.precio_peso = precio_peso;
 	}
-	public int getPrecio_dolar() {
+	public float getPrecio_dolar() {
 		return precio_dolar;
 	}
-	public void setPrecio_dolar(int precio_dolar) {
+	public void setPrecio_dolar(float precio_dolar) {
 		this.precio_dolar = precio_dolar;
 	}
-	public int getPrecio_euro() {
+	public float getPrecio_euro() {
 		return precio_euro;
 	}
-	public void setPrecio_euro(int precio_euro) {
+	public void setPrecio_euro(float precio_euro) {
 		this.precio_euro = precio_euro;
 	}
 }
